+++ cucumber-core/src/main/java/io/cucumber/core/runner/TestCase.java
@@ -24,7 +24,7 @@ import java.util.stream.Collectors;
 
 import static io.cucumber.core.runner.ExecutionMode.DRY_RUN;
 import static io.cucumber.core.runner.ExecutionMode.RUN;
-import static io.cucumber.messages.TimeConversion.javaInstantToTimestamp;
+import static io.cucumber.messages.Convertor.toMessage;
 import static java.util.Collections.singletonList;
 import static java.util.stream.Collectors.toList;
 
@@ -208,7 +208,8 @@ final class TestCase implements io.cucumber.plugin.event.TestCase {
             0L,
             executionId.toString(),
             id.toString(),
-            javaInstantToTimestamp(start)));
+            Thread.currentThread().getName(),
+            toMessage(start)));
         bus.send(envelope);
     }
 
@@ -217,7 +218,7 @@ final class TestCase implements io.cucumber.plugin.event.TestCase {
     ) {
         bus.send(new TestCaseFinished(stop, this, result));
         Envelope envelope = Envelope.of(new io.cucumber.messages.types.TestCaseFinished(executionId.toString(),
-            javaInstantToTimestamp(stop), false));
+            toMessage(stop), false));
         bus.send(envelope);
     }