+++ cucumber-spring/src/main/java/io/cucumber/spring/TestContextAdaptor.java
@@ -13,41 +13,51 @@ import java.lang.reflect.Method;
 import java.util.ArrayDeque;
 import java.util.Collection;
 import java.util.Deque;
+import java.util.function.Supplier;
 
 import static io.cucumber.spring.CucumberTestContext.SCOPE_CUCUMBER_GLUE;
-import static org.springframework.beans.factory.config.AutowireCapableBeanFactory.AUTOWIRE_NO;
+import static org.springframework.beans.factory.config.AutowireCapableBeanFactory.AUTOWIRE_CONSTRUCTOR;
 
 class TestContextAdaptor {
 
     private static final Object monitor = new Object();
-
     private final TestContextManager delegate;
     private final ConfigurableApplicationContext applicationContext;
-    private final Collection<Class<?>> glueClasses;
     private final Deque<Runnable> stopInvocations = new ArrayDeque<>();
     private Object delegateTestInstance;
 
-    TestContextAdaptor(
-            TestContextManager delegate,
+    static TestContextAdaptor create(
+            Supplier<TestContextManager> testContextManagerSupplier,
             Collection<Class<?>> glueClasses
     ) {
-        TestContext testContext = delegate.getTestContext();
-        ConfigurableApplicationContext applicationContext = (ConfigurableApplicationContext) testContext
-                .getApplicationContext();
-        this.delegate = delegate;
-        this.applicationContext = applicationContext;
-        this.glueClasses = glueClasses;
+        synchronized (monitor) {
+            // While under construction, the TestContextManager delegate will
+            // build a cached version of the application context configuration.
+            // Since Spring Boot 3 and in combination with AOT building this
+            // configuration is not idempotent (#2686).
+            TestContextManager delegate = testContextManagerSupplier.get();
+
+            TestContext testContext = delegate.getTestContext();
+            ConfigurableApplicationContext applicationContext = (ConfigurableApplicationContext) testContext
+                    .getApplicationContext();
+
+            // The TestContextManager delegate makes the application context
+            // available to other threads. Registering the glue however modifies
+            // the application context. To avoid concurrent modification issues
+            // (#1823, #1153, #1148, #1106) we do this serially.
+            registerGlueCodeScope(applicationContext);
+            registerStepClassBeanDefinitions(applicationContext.getBeanFactory(), glueClasses);
+
+            return new TestContextAdaptor(delegate);
+        }
     }
 
-    public final void start() {
-        // The TestContextManager delegate makes the application context
-        // available to other threads. Registering the glue however modifies the
-        // application context. To avoid concurrent modification issues (#1823,
-        // #1153, #1148, #1106) we do this serially.
-        synchronized (monitor) {
-            registerGlueCodeScope(applicationContext);
-            registerStepClassBeanDefinitions(applicationContext.getBeanFactory());
-        }
+    TestContextAdaptor(TestContextManager delegate) {
+        this.delegate = delegate;
+        this.applicationContext = (ConfigurableApplicationContext) delegate.getTestContext().getApplicationContext();
+    }
+
+    final void start() {
         stopInvocations.push(this::notifyTestContextManagerAboutAfterTestClass);
         notifyContextManagerAboutBeforeTestClass();
         stopInvocations.push(this::stopCucumberTestContext);
@@ -92,11 +102,25 @@ class TestContextAdaptor {
         // using their default constructor and now allow them to be injected
         // into other step definition classes.
         try {
-            Class<?> delegateTestClass = delegate.getTestContext().getTestClass();
-            Object delegateTestInstance = applicationContext.getBeanFactory().autowire(delegateTestClass, AUTOWIRE_NO,
-                false);
-            delegate.prepareTestInstance(delegateTestInstance);
-            this.delegateTestInstance = delegateTestInstance;
+            Class<?> beanClass = delegate.getTestContext().getTestClass();
+
+            ConfigurableListableBeanFactory beanFactory = applicationContext.getBeanFactory();
+            // Note: By providing AUTOWIRE_CONSTRUCTOR the
+            // AbstractAutowireCapableBeanFactory does not invoke
+            // 'populateBean' and effectively creates a raw bean.
+            Object bean = beanFactory.autowire(beanClass, AUTOWIRE_CONSTRUCTOR, false);
+
+            // But it works out well for us. Because now the
+            // DependencyInjectionTestExecutionListener will invoke
+            // 'autowireBeanProperties' which will populate the bean.
+            delegate.prepareTestInstance(bean);
+
+            // Because the bean is created by a factory, it is not added to
+            // the application context yet.
+            CucumberTestContext scenarioScope = CucumberTestContext.getInstance();
+            scenarioScope.put(beanClass.getName(), bean);
+
+            this.delegateTestInstance = bean;
         } catch (Exception e) {
             throw new CucumberBackendException(e.getMessage(), e);
         }
@@ -111,7 +135,7 @@ class TestContextAdaptor {
         }
     }
 
-    private void registerGlueCodeScope(ConfigurableApplicationContext context) {
+    private static void registerGlueCodeScope(ConfigurableApplicationContext context) {
         while (context != null) {
             ConfigurableListableBeanFactory beanFactory = context.getBeanFactory();
             // Scenario scope may have already been registered by another
@@ -132,14 +156,16 @@ class TestContextAdaptor {
         }
     }
 
-    private void registerStepClassBeanDefinitions(ConfigurableListableBeanFactory beanFactory) {
+    private static void registerStepClassBeanDefinitions(
+            ConfigurableListableBeanFactory beanFactory, Collection<Class<?>> glueClasses
+    ) {
         BeanDefinitionRegistry registry = (BeanDefinitionRegistry) beanFactory;
         for (Class<?> glue : glueClasses) {
             registerStepClassBeanDefinition(registry, glue);
         }
     }
 
-    private void registerStepClassBeanDefinition(BeanDefinitionRegistry registry, Class<?> glueClass) {
+    private static void registerStepClassBeanDefinition(BeanDefinitionRegistry registry, Class<?> glueClass) {
         String beanName = glueClass.getName();
         // Step definition may have already been
         // registered as a bean by another thread.
@@ -152,7 +178,7 @@ class TestContextAdaptor {
                 .getBeanDefinition());
     }
 
-    public final void stop() {
+    final void stop() {
         // Cucumber only supports 1 set of before/after semantics while JUnit
         // and Spring have 2 sets. So here we use a stack to ensure we don't
         // invoke only the matching after methods for each before methods.